+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 2.5.4 - 2018-07-29
 
 ### Changed
-- The `generateTransformsBeforePageLoad` Craft config setting is now automatically disabled for all Element API endpoints. ([#81](https://github.com/craftcms/element-api/issues/81))
+- The `generateTransformsBeforePageLoad` Craft config setting is now automatically enabled for all Element API endpoints. ([#81](https://github.com/craftcms/element-api/issues/81))
 
 ### Fixed
 - Fixed a bug where it wasn’t possible to set `cache`, `serializer`, `jsonOptions`, `pretty`, `includes`, and `excludes` endpoint config options in the `defaults` array. ([#69](https://github.com/craftcms/element-api/pull/69))