+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Changed
+- Updated Fractal to 0.18. ([#109](https://github.com/craftcms/element-api/issues/109))
+
 ### Fixed
 - Fixed a bug where field types that return objects were getting converted to arrays by the default element transformer. Now their [serialized value](https://docs.craftcms.com/api/v3/craft-base-fieldinterface.html#method-serializevalue) is used instead. ([#75](https://github.com/craftcms/element-api/issues/75))